+++ packages/x-adapter-platform/src/types/responses/related-prompts-response.model.ts
@@ -1,0 +1,11 @@
+import { PlatformRelatedPrompt } from '../models/related-prompt.model';
+
+/**
+ * Response for the `related prompts` endpoint.
+ */
+export interface PlatformRelatedPromptsResponse {
+  data: {
+    relatedprompts: PlatformRelatedPrompt[];
+  };
+  status: number;
+}